More work for consistent use of paths.

+++ managers/FileManager.py
@@ -17,12 +17,12 @@ from managers.Request import Request, ExporterFiles, ExporterRange
 
 class FileManager(abc.ABC):
     def __init__(self, exporter_files: ExporterFiles, game_id, data_dir: str, date_range: Dict[str,datetime]):
-        self._file_names : Dict[str,Union[str,None]] = {"sessions_f":None, "events_f":None}
-        self._zip_names  : Dict[str,Union[str,None]] = {"sessions_f":None, "events_f":None}
+        self._file_names : Dict[str,Union[Path,None]] = {"sessions_f":None, "events_f":None}
+        self._zip_names  : Dict[str,Union[Path,None]] = {"sessions_f":None, "events_f":None}
         self._files      : Dict = {"sessions_f":None, "events_f":None}
-        self._data_dir   : str  = data_dir
+        self._data_dir   : Path  = Path(data_dir)
         self._game_id    : str  = game_id
-        self._readme_path: str
+        self._readme_path: Path
         self._dataset_id : str
         self._short_hash : str
         try:
@@ -34,14 +34,14 @@ class FileManager(abc.ABC):
             repo = git.Repo(search_parent_directories=True)
             self._short_hash = repo.git.rev_parse(repo.head.object.hexsha, short=7)
             # then set up our paths.
-            full_data_dir = self._data_dir + game_id
-            self._readme_path = f"{full_data_dir}/readme.md"
-            base_path = f"{full_data_dir}/{self._dataset_id}_{self._short_hash}"
+            full_data_dir     : Path = self._data_dir / game_id
+            self._readme_path : Path = full_data_dir / "readme.md"
+            base_path         : Path = full_data_dir / f"{self._dataset_id}_{self._short_hash}"
             # finally, generate file names.
-            self._file_names["sessions_f"] = base_path+"_session-features.csv" if exporter_files.sessions else None
-            self._file_names["events_f"] = base_path+"_events.tsv" if exporter_files.events else None
-            self._zip_names["sessions_f"] = base_path+"_session-features.zip" if exporter_files.sessions else None
-            self._zip_names["events_f"] = base_path+"_events.zip" if exporter_files.events else None
+            self._file_names["sessions_f"] = base_path / "_session-features.csv" if exporter_files.sessions else None
+            self._file_names["events_f"]   = base_path / "_events.tsv" if exporter_files.events else None
+            self._zip_names["sessions_f"]  = base_path / "_session-features.zip" if exporter_files.sessions else None
+            self._zip_names["events_f"]    = base_path / "_events.zip" if exporter_files.events else None
         except Exception as err:
             msg = f"{type(err)} {str(err)}"
             utils.Logger.Log(msg, logging.ERROR)
@@ -58,11 +58,12 @@ class FileManager(abc.ABC):
 
     def OpenFiles(self):
         # Ensure we have a data directory.
-        full_data_dir = Path(self._data_dir) / self._game_id
+        self._data_dir.mkdir(exist_ok=True)
+        full_data_dir : Path = self._data_dir / self._game_id
         full_data_dir.mkdir(exist_ok=True)
         # Then open up the files themselves.
         self._files["sessions_f"] = open(self._file_names["sessions_f"], "w", encoding="utf-8") if (self._file_names["sessions_f"] is not None) else None
-        self._files["events_f"] = open(self._file_names["events_f"], "w", encoding="utf-8") if (self._file_names["events_f"] is not None) else None
+        self._files["events_f"]   = open(self._file_names["events_f"],   "w", encoding="utf-8") if (self._file_names["events_f"] is not None) else None
 
     def CloseFiles(self):
         if self._files["sessions_f"] is not None:
@@ -72,7 +73,7 @@ class FileManager(abc.ABC):
 
     def ZipFiles(self):
         try:
-            existing_csvs = utils.loadJSONFile("file_list.json", self._data_dir)
+            existing_csvs = utils.loadJSONFile("file_list.json", str(self._data_dir))
         except Exception as err:
             existing_csvs = {}
         # if we have already done this dataset before, rename old zip files
@@ -82,9 +83,9 @@ class FileManager(abc.ABC):
                 src_sessions_f = existing_csvs[self._game_id][self._dataset_id]['sessions_f']
                 src_events_f = existing_csvs[self._game_id][self._dataset_id]['events_f']
                 if src_sessions_f is not None and self._zip_names["sessions_f"] is not None:
-                    os.rename(src_sessions_f, self._zip_names["sessions_f"])
+                    os.rename(src_sessions_f, str(self._zip_names["sessions_f"]))
                 if src_events_f is not None and self._zip_names["events_f"] is not None:
-                    os.rename(src_events_f, self._zip_names["events_f"])
+                    os.rename(src_events_f, str(self._zip_names["events_f"]))
         except Exception as err:
             msg = f"Error while setting up zip files! {type(err)} : {err}"
             utils.Logger.Log(msg, logging.ERROR)
@@ -97,7 +98,8 @@ class FileManager(abc.ABC):
                 sessions_zip_file = zipfile.ZipFile(self._zip_names["sessions_f"], "w", compression=zipfile.ZIP_DEFLATED)
                 self._addToZip(path=self._file_names["sessions_f"], zip_file=sessions_zip_file, path_in_zip=f"{base_path}_session-features.csv")
                 self._addToZip(path=self._readme_path,        zip_file=sessions_zip_file, path_in_zip=f"{self._dataset_id}/readme.md")
-                os.remove(self._file_names["sessions_f"])
+                if self._file_names["sessions_f"] is not None:
+                    os.remove(self._file_names["sessions_f"])
             except FileNotFoundError as err:
                 utils.Logger.Log(f"FileNotFoundError Exception: {err}", logging.ERROR)
                 traceback.print_tb(err.__traceback__)
@@ -108,7 +110,8 @@ class FileManager(abc.ABC):
                 events_zip_file = zipfile.ZipFile(self._zip_names["events_f"], "w", compression=zipfile.ZIP_DEFLATED)
                 self._addToZip(path=self._file_names["events_f"], zip_file=events_zip_file, path_in_zip=f"{base_path}_events.tsv")
                 self._addToZip(path=self._readme_path,        zip_file=events_zip_file, path_in_zip=f"{self._dataset_id}/readme.md")
-                os.remove(self._file_names["events_f"])
+                if self._file_names["events_f"] is not None:
+                    os.remove(self._file_names["events_f"])
             except FileNotFoundError as err:
                 utils.Logger.Log(f"FileNotFoundError Exception: {err}", logging.ERROR)
                 traceback.print_tb(err.__traceback__)
@@ -129,7 +132,7 @@ class FileManager(abc.ABC):
     #  @param num_sess      The number of sessions included in the recent export.
     def WriteMetadataFile(self, date_range: Dict[str,datetime], num_sess: int):
         # First, ensure we have a data directory.
-        full_data_dir = Path(self._data_dir) / self._game_id
+        full_data_dir : Path = self._data_dir / self._game_id
         try:
             full_data_dir.mkdir(exist_ok=True)
         except Exception as err:
@@ -159,14 +162,14 @@ class FileManager(abc.ABC):
             else:
                 metadata  = \
                 {
-                    "game_id":self._game_id,
+                    "game_id"      :self._game_id,
                     "dataset_id"   :self._dataset_id,
-                    "sessions_f":self._zip_names["sessions_f"],
-                    "events_f":self._zip_names["events_f"],
+                    "sessions_f"   :str(self._zip_names["sessions_f"]),
+                    "events_f"     :str(self._zip_names["events_f"]),
                     "start_date"   :date_range['min'].strftime("%m/%d/%Y"),
                     "end_date"     :date_range['max'].strftime("%m/%d/%Y"),
                     "date_modified":datetime.now().strftime("%m/%d/%Y"),
-                    "sessions":num_sess
+                    "sessions"     :num_sess
                 }
                 meta_file.write(json.dumps(metadata, indent=4))
                 meta_file.close()
@@ -181,36 +184,36 @@ class FileManager(abc.ABC):
         self._backupFileExportList()
         existing_csvs = {}
         try:
-            existing_csvs = utils.loadJSONFile("file_list.json", self._data_dir)
+            existing_csvs = utils.loadJSONFile("file_list.json", str(self._data_dir))
         except FileNotFoundError as err:
             utils.Logger.toFile("file_list.json does not exist.", logging.WARNING)
         except Exception as err:
             msg = f"Could not load file list. {type(err)} {str(err)}"
             utils.Logger.toFile(msg, logging.ERROR)
         finally:
-            with open(f"{self._data_dir}file_list.json", "w") as existing_csv_file:
+            with open(self._data_dir / "file_list.json", "w") as existing_csv_file:
                 utils.Logger.toStdOut(f"opened csv file at {existing_csv_file.name}", logging.INFO)
                 if not self._game_id in existing_csvs.keys():
                     existing_csvs[self._game_id] = {}
                 # sessions_stat = os.stat(sessions_csv_full_path)
                 prior_export = self._dataset_id in existing_csvs[self._game_id].keys()
-                sessions_path = self._zip_names["sessions_f"] if self._zip_names["sessions_f"] is not None else (existing_csvs[self._game_id][self._dataset_id]["sessions_f"] if prior_export else None)
-                events_path = self._zip_names["events_f"] if self._zip_names["events_f"] is not None else (existing_csvs[self._game_id][self._dataset_id]["events_f"] if prior_export else None)
+                sessions_path = str(self._zip_names["sessions_f"]) if self._zip_names["sessions_f"] is not None else (existing_csvs[self._game_id][self._dataset_id]["sessions_f"] if prior_export else None)
+                events_path   = str(self._zip_names["events_f"]) if self._zip_names["events_f"] is not None else (existing_csvs[self._game_id][self._dataset_id]["events_f"] if prior_export else None)
                 existing_csvs[self._game_id][self._dataset_id] = \
                 {
-                    "sessions_f":sessions_path,
-                    "events_f":events_path,
+                    "sessions_f"   :sessions_path,
+                    "events_f"     :events_path,
                     "start_date"   :date_range['min'].strftime("%m/%d/%Y"),
                     "end_date"     :date_range['max'].strftime("%m/%d/%Y"),
                     "date_modified":datetime.now().strftime("%m/%d/%Y"),
-                    "sessions":num_sess
+                    "sessions"     :num_sess
                 }
                 existing_csv_file.write(json.dumps(existing_csvs, indent=4))
 
     def _backupFileExportList(self) -> bool:
         try:
-            src = Path(self._data_dir) / "file_list.json"
-            dest = Path(self._data_dir) / "file_list.json.bak"
+            src  : Path = self._data_dir / "file_list.json"
+            dest : Path = self._data_dir / "file_list.json.bak"
             if src.exists():
                 shutil.copyfile(src=src, dst=dest)
             else:
